Convert to TypeScript
In order to align this repo with our other repos, this commit converts
the existing JavaScript to TypeScript, and adds the dependencies and
configuration to support this. It also adds `typedoc` and updates the
existing GitHub workflows to publish documentation. This means we no
longer need to explicitly ship type definitions as they will be
generated automatically.

Finally, there were a couple of standardization-related things that were
missed in previous updates, so this adds those as well. After this
commit this repo should be fully compliant with the module template.